Dispose Enumerators if possible
In some cases it is illegal to have multiple enumerators open on the same
enumerable at the same time. This is the case with EntityFramework results
which come from a database that doesn't support concurrent DbDataReaders
(e.g. PostgreSQL).

When comparing enumerables, added a check if the enumerators are disposable,
and if so, dispose them.