Allow AspNet assemblies to load correctly when treated as platform
Previously, xunit assemblies were loaded by simple name, using
reflection with an AssemblyName without a version set.  This works
when the assembly is not treated as a trusted platform assembly.
When the assembly is treated as a trusted platform assembly, the bind
will fail with FUSION_E_REF_DEF_MISMATCH since binds to platform
assemblies must include a version.  By adding Version(0, 0) to the
AssemblyName, the bind will succeed because the assembly def version
will be greater than the assembly ref version.